pyproject.toml
[project]
name = "fastapi-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.128.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.46",
]
